Update manifest when deleting state from a non-empty bundle
This fixes a bug where the manifest would contain dangling references to old
state if multiple states were saved to a single bundle and one is deleted.

As far as I know this has not been encountered by anyone in reality, since this
is an uncommon scenario (it's somewhat in question whether it should be
supported at all).