Use standard assert for test expectations
The old one aborted anyway, and I've been using assert for tests in other
projects for a while now, and never really missed having something more
complex.  The lack of a dependency or repetitive code is nice.

Since the unit tests are about to be split up, the abort thing will become less
of an issue anyway.