Use the canonical path for state directories
This fixes an issue on MacOS if hosts use a temporary directory generated with
mkdtemp, which can create paths that involve symlinks.  This is problematic
since the parent directory comparisons here need to determine if a file is
really in one of the state directories.  The actual paths here shouldn't be
visible in saved state, so resolve them fully so that the directory membership
tests work correctly and copies and/or links are created when necessary.