Remove project and version number from man page OS field
The mandoc documentation suggests leaving this blank.  Although I still think
it's nicer for third-party projects to put their own versions here (as many GNU
utilities do, for example), it's a hassle in practice to maintain the version
over time, and I'm not sure what to do about the date in that case.