Add node hash utility and replace lilv_world_filter_model()
This pattern is used where a separate structure of nodes is needed (generally
when modifying the model, so iterators will be invalided).  The old code was
abusing models as expensive lists for this purpose, which is both inefficient
and tedious.  Replace this with a hash of node pointers which retains
references to nodes (in case the last instance of a node in a model is removed
while the hash it still alive).