Simplify lilv_world_get() implementation and clarify documentation
For consistency with lilv_world_find_nodes(), and towards the grander goal of
reducing the amount of indexing required in general.

This might sort of work with a wildcard predicate, but since there's no index
for that, it will be very slow at best.  So, keep the implementation oblivious
but strengthen the documentation to only support a subject or object wildcard.
This could be considered breakage, but any host actually doing this should
address it anyway.